- moved BinarySearchTreeTest.java to the tight package
